fix: accommodate u64 share accounting and point to correct stratum branch

+++ miner-apps/jd-client/src/lib/channel_manager/downstream_message_handler.rs
@@ -989,7 +989,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                                 channel_id,
                                 last_sequence_number: share_accounting.get_last_share_sequence_number(),
                                 new_submits_accepted_count: share_accounting.get_last_batch_accepted(),
-                                new_shares_sum: share_accounting.get_last_batch_work_sum() as u64,
+                                new_shares_sum: share_accounting.get_last_batch_work_sum(),
                             };
                             info!("SubmitSharesStandard on downstream channel: {} ✅", success);
                             messages.push((downstream.downstream_id, Mining::SubmitSharesSuccess(success)).into());
@@ -1023,7 +1023,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                             channel_id,
                             last_sequence_number: share_accounting.get_last_share_sequence_number(),
                             new_submits_accepted_count: share_accounting.get_last_batch_accepted(),
-                            new_shares_sum: share_accounting.get_last_batch_work_sum() as u64,
+                            new_shares_sum: share_accounting.get_last_batch_work_sum(),
                         };
                         messages.push((
                             downstream.downstream_id,
@@ -1240,7 +1240,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                                 channel_id,
                                 last_sequence_number: share_accounting.get_last_share_sequence_number(),
                                 new_submits_accepted_count: share_accounting.get_last_batch_accepted(),
-                                new_shares_sum: share_accounting.get_last_batch_work_sum() as u64,
+                                new_shares_sum: share_accounting.get_last_batch_work_sum(),
                             };
                             info!("SubmitSharesExtended on downstream channel: {} ✅", success);
                             messages.push((downstream.downstream_id, Mining::SubmitSharesSuccess(success)).into());
@@ -1272,7 +1272,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                             channel_id,
                             last_sequence_number: share_accounting.get_last_share_sequence_number(),
                             new_submits_accepted_count: share_accounting.get_last_batch_accepted(),
-                            new_shares_sum: share_accounting.get_last_batch_work_sum() as u64,
+                            new_shares_sum: share_accounting.get_last_batch_work_sum(),
                         };
                         is_downstream_share_valid = true;
                         messages.push((
